requirements.txt
setuptools==39.0.1
click
grpcio==1.15.0
grpcio-tools==1.15.0
grpcio-health-checking==1.15.0
protobuf>=3.6.1
googleapis-common-protos==1.6.0
requests==2.22.0
Flask==1.1.1
python-magic>=0.4.14
